+++ src/People.py
@@ -31,7 +31,7 @@ class Student(Person):
 			self.mark = round(mean(self.assignments.values()), 1)
 	
 	# "Local" student implementation of the add assignment function where details such as assignment grades can be adjusted for each student
-	def add_assignment(self, assignment_name: str, assignment_mark: int) -> None:
+	def add_assignment(self, assignment_name: str, assignment_mark: float) -> None:
 		self.assignments[assignment_name] = assignment_mark
 		self.average()
